[Rename] Store 데이터 함수명 및 파일명 변경
