+++ components/StoreSchedule.tsx
@@ -1,6 +1,6 @@
 import { StyleSheet } from "react-native";
 import { View } from "react-native";
-import { StoreSchedulesType } from "../API/getCafeDetail";
+import { StoreSchedulesType } from "../API/getStoreDetail";
 import { dayOfTheWeek } from "../function/dayOfTheWeek";
 import CustomText from "./CustomText";
 interface Props {
